backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Authentication
PyJWT==2.8.0

# Database
psycopg2-binary==2.9.9

# WebRTC (Real-time video streaming)
aiortc==1.9.0
aiohttp>=3.9.0
av>=12.0.0

# AI / Computer Vision
opencv-python==4.9.0.80
numpy==1.26.3
torch>=2.0.0
ultralytics==8.3.171
insightface==0.7.3

# Inference Optimization (using CPU version to avoid CUDA loading errors)
# Note: YOLO uses PyTorch's native CUDA, not ONNX Runtime
onnxruntime==1.17.0

# Utilities
pandas==2.1.1
pyserial
PyJWT==2.8.0